# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sudosos",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/storages/owner/{owner_id}": {
            "get": {
                "tags": [
                    "storage"
                ],
                "summary": "Returns all storages for an organ.",
                "description": "Returns all storages for an organ.",
                "operationId": "getAllStorages",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "owner_id",
                        "in": "path",
                        "description": "Id of the owner",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Owner not found"
                    }
                }
            }
        },
        "/storages": {
            "post": {
                "tags": [
                    "storage"
                ],
                "summary": "Create a new storage.",
                "description": "Create a new storage.",
                "operationId": "createStorage",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "storage",
                        "in": "body",
                        "description": "Model of the storage to store",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inputStorage"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Storage invalid."
                    }
                }
            }
        },
        "/storages/{id}": {
            "get": {
                "tags": [
                    "storage"
                ],
                "summary": "Returns storage by id.",
                "description": "Returns a storage with a specified id.",
                "operationId": "getStorage",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the storage",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Storage not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "storage"
                ],
                "summary": "Updates a storage by id.",
                "description": "Updates the storage.",
                "operationId": "updateStorage",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "request",
                        "in": "path",
                        "description": "Request body in JSON.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the storage",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Storage succesfully deleted"
                    },
                    "400": {
                        "description": "Storage not valid"
                    },
                    "404": {
                        "description": "Storage not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "storage"
                ],
                "summary": "Delete a storage by id.",
                "description": "Delete a storage by id.",
                "operationId": "deleteStorage",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the storage",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Storage succesfully deleted."
                    },
                    "404": {
                        "description": "Storage not found."
                    }
                }
            }
        },
        "/storages/{id}/reinstate": {
            "put": {
                "tags": [
                    "storage"
                ],
                "summary": "Reinstate a storage by id.",
                "description": "Reinstate a storage by id.",
                "operationId": "reinstateStorage",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the storage",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Storage succesfully reinstated"
                    },
                    "404": {
                        "description": "Storage not found"
                    },
                    "409": {
                        "description": "Storage already active."
                    }
                }
            }
        },
        "/storages/{id}/{property}": {
            "get": {
                "tags": [
                    "storage"
                ],
                "summary": "Returns a property of a storage by id.",
                "description": "Returns a property of a storage by id.",
                "operationId": "getStorageProperty",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the storage",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "property",
                        "in": "path",
                        "description": "Property of the storage",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Storage or Property not found, see message body"
                    }
                }
            },
            "put": {
                "tags": [
                    "storage"
                ],
                "summary": "Update a property of a storage by id.",
                "description": "Update a property of a storage by id.",
                "operationId": "getStorageProperty",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the storage",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "request",
                        "in": "path",
                        "description": "Request body in JSON",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "property",
                        "in": "path",
                        "description": "Property of the storage",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Storage succesfully updated"
                    },
                    "404": {
                        "description": "Storage or Property not found, see message body"
                    },
                    "400": {
                        "description": "Invalid property value"
                    }
                }
            }
        },
        "/storages/{id}/stores": {
            "get": {
                "tags": [
                    "storage"
                ],
                "summary": "Returns all products in a storage by id.",
                "description": "Returns all products in a storage by id.",
                "operationId": "getStorageProducts",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the storage",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Storage not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "inputStorage": {
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "name": {
                    "description": "The name of the storage",
                    "type": "string"
                },
                "owner_id": {
                    "description": "ID of the owner from this storage",
                    "type": "string"
                }
            }
        },
...
    }
}